+++ webknossos/webknossos/client/_resolve_short_link.py
@@ -2,8 +2,7 @@ import logging
 import re
 from urllib.parse import urlparse
 
-from ._generated.api.default import short_link_by_key
-from .context import _get_generated_client, webknossos_context
+from .context import _get_api_client, webknossos_context
 
 logger = logging.getLogger(__name__)
 
@@ -26,13 +25,8 @@ def resolve_short_link(url: str) -> str:
         webknossos_url = f"{parts.scheme}://{parts.netloc}"
         short_link_key = match.group("short_link_key")
         with webknossos_context(url=webknossos_url):
-            client = _get_generated_client()
-            response = short_link_by_key.sync(
-                key=short_link_key,
-                client=client,
-            )
-        assert response is not None, f"Could not resolve short link {url}."
-        return response.long_link
+            client = _get_api_client()
+            return client.short_link_by_key(key=short_link_key).long_link
     except Exception as e:
         logger.warning(f"Got an error during short link resolution of link {url}: {e}")
         return url